Move .AsOption into GlobalExtensions.
